+++ rasgoql/rasgoql/primitives/transforms.py
@@ -4,7 +4,7 @@ Primitive Classes
 from __future__ import annotations
 from collections import OrderedDict
 import logging
-from typing import Callable, List, TYPE_CHECKING
+from typing import Any, Callable, Dict, List, TYPE_CHECKING, TypeVar, Union
 
 import pandas as pd
 import rasgotransforms as rtx
@@ -83,7 +83,7 @@ class TransformableClass:
     def transform(
         self,
         name: str,
-        arguments: OrderedDict = None,
+        arguments: Dict[str, Union[Any, Dataset, SQLChain]] = None,
         output_alias: str = None,
         **kwargs,
     ) -> SQLChain:
@@ -109,7 +109,7 @@ class TransformableClass:
         arguments = arguments if arguments else {}
         arguments.update(kwargs)
         arguments['source_table'] = source_table
-        new_transform = Transform(name, arguments, namespace, output_alias, self._dw)
+        new_transform = Transform(name, flatten_transform_arguments(arguments), namespace, output_alias, self._dw)
         transforms.append(new_transform)
         return SQLChain(entry_table, namespace, transforms, self._dw)
 
@@ -311,7 +311,7 @@ class SQLChain(TransformableClass):
 
         NOTE: This property will be dynamic until the Chain is finally saved
         """
-        return self._dw.get_schema(self.fqtn, self.sql())
+        return self._dw.get_schema(self.sql())
 
     @property
     def output_table(self) -> Dataset:
@@ -430,3 +430,24 @@ class SQLChain(TransformableClass):
             response='df',
             batches=batches,
         )
+
+
+S = TypeVar("S")
+
+
+def flatten_transform_arguments(arguments: Dict[str, Union[S, Dataset, SQLChain]]) -> Dict[str, Union[S, str]]:
+    """
+    Looks for primitive classes in a list of transform arguments and
+    flattens them into renderable SQL args
+    """
+    arguments_out = {}
+    for k, v in arguments.items():
+        # If Dataset -> set to fqtn
+        if isinstance(v, Dataset):
+            arguments_out[k] = v.fqtn
+        # If SQLChain -> render as cte
+        elif isinstance(v, SQLChain):
+            arguments_out[k] = f"({v.sql()})"
+        else:
+            arguments_out[k] = v
+    return arguments_out